requirements.txt
azure-core==1.25.1
azure-functions==1.11.2
azure-storage-blob==12.13.1
pandas==2.0.2
pydantic==2.5.2
pyodbc==4.0.34
SQLAlchemy==1.4.40
requests==2.31.0
requests-oauthlib==1.3.1
strictyaml==1.7.3